% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECIbootstrapTest.R
\name{BCa}
\alias{BCa}
\title{Bootstrap p-value}
\usage{
BCa(theta, theta_hat, conf.level = 0.95)
}
\arguments{
\item{theta}{A named vector of bootstrapped ECI values}

\item{theta_hat}{The original ECI value}

\item{conf.level}{The confidence level for the confidence interval.}
}
\value{
The BCa confidence interval
}
\description{
This function calculates the BCa confidence interval. The function is adapted from the bca function of the coxed package.
}
