% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECIbootstrapTest.R
\name{bootstrap.pval}
\alias{bootstrap.pval}
\title{Bootstrap p-value}
\usage{
bootstrap.pval(data, ES)
}
\arguments{
\item{data}{A named vector of bootstrapped ECI values}

\item{ES}{the original ECI value}
}
\value{
The p-value based on BCa confidence interval
}
\description{
This function calculates an approximate p-value for a bootstrapped value based on BCa confidence interval.
}
