% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECIbootstrapTest.R
\name{bootstrap.qval}
\alias{bootstrap.qval}
\title{Bootstrap q-value}
\usage{
bootstrap.qval(pval)
}
\arguments{
\item{pval}{A named vector of p-values}
}
\value{
A vector of q-values
}
\description{
This function calculates q-values for p-values from multiple ECI bootstrap tests
}
