% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpr.R
\name{diffExpr}
\alias{diffExpr}
\title{Differential Expression using limma}
\usage{
diffExpr(data, targets)
}
\arguments{
\item{data}{A gene expression matrix with genes as rows and subjects as columns.}

\item{targets}{A data.frame of subject "ID" and "Type" as columns. "Type" has to be "normal" or "tumor".}
}
\value{
A matrix of differential gene expression values in log2FC format, standard deviation,
p value, and degrees of freedom.
}
\description{
This function performs differential expression using Limma and
calculates next to the pvalue the standard deviation.
}
