% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECIbootstrapTest.R
\name{ECIbootstrapTest}
\alias{ECIbootstrapTest}
\title{ECI bootstrap test}
\usage{
ECIbootstrapTest(data1, data2, targets1, targets2, filter = TRUE, alpha = 0.05)
}
\arguments{
\item{targets1}{A data.frame of subject "ID" and "Type" as columns. "Type" has to be "normal" or "tumor".}

\item{targets2}{A data.frame of subject "ID" and "Type" as columns. "Type" has to be "normal" or "tumor".}

\item{filter}{A boolean parameter indicating if the returned values should by filtered by genes expressing}

\item{alpha}{Significance level for the confidence interval.
being differentially expressed in at least one of the studies.}

\item{geneExpr1}{A gene expression matrix with genes as rows and subjects as columns.}

\item{geneExpr2}{A gene expression matrix with genes as rows and subjects as columns.}
}
\value{
A matrix of ECI values, confidence interval, p value, and q value.
}
\description{
This function performs a molecular equivalence test based on the ECI statistic using a bootstrapt test.
}
